// src/app/company/models/company-detail.model.ts
export class CompanyDetail {
    listName: string;
    tickerName: string;
    summary: string;
    open: string;
    high: string;
    marketCap: string;
    peRatio: string;
    low: string;
    weekHigh52: string;
    weekLow52: string;
    prevClose: string;
    yield: string;
    trend: string;
}